Update exrtma3d_gsianl.ksh
This script was adjusted to follow the logic posed in the operational HRRR gsianl exscript. The key change is that the number of gdas ensemble is always counted to provide a value for the variable "nummem" regardless if there are a sufficient number of hrrrdas files or not.  After testing over several cycles with and without 36 hrrrdas files, the now runs correctly in either case with it defaulting to gdas if all 36 hrrrdas ensemble file elements are not present.