Resolution for late-arriving hrrrdas file within the gsianl ex-script (#47)
* Resolved the late-arriving hrrrdas file logic within the gsianl ex-script and added stdout and OUTPUT file copy commands needed for archiving.

NOTE: Since the development machine has reverted back to the 35+ minute obs arrival time, the HRRRDAS_CUTOFF set in rtma3d_pbspro_conus.xml should be set to at least 40 minutes only on parallels set up on WCOSS development.

* Update exrtma3d_gsianl.ksh

* Update exrtma3d_gsianl.ksh

This script was adjusted to follow the logic posed in the operational HRRR gsianl exscript. The key change is that the number of gdas ensemble is always counted to provide a value for the variable "nummem" regardless if there are a sufficient number of hrrrdas files or not.  After testing over several cycles with and without 36 hrrrdas files, the now runs correctly in either case with it defaulting to gdas if all 36 hrrrdas ensemble file elements are not present.