Added a scrubbing step to the existing 3DRTMA workflow. This step is invoked at the start and clears COM and DATA directories that exceed their respective retention periods. (#69)
* Added a scrubbing step to the existing 3DRTMA workflow. This step is invoked at the start and clears COM and DATA directories that exceed their respective retention periods.

* Update rtma3d_pbspro_conus.xml

Resolved resource allocation for clean step.

* Update rtma3d_pbspro_alaska.xml

Resolved resource allocation for clean step.

* Update exrtma3d_clean.ksh

Removed --login argument in environment call as its use is discouraged by NCO.

* Update rtma3d_pbspro_conus.xml

Increased the COM retention period to 5 days for autoQC requirements.

* Update rtma3d_pbspro_alaska.xml

Increased the COM retention period to 5 days for autoQC requirements.

* Update exrtma3d_clean.ksh

A couple of additional changes to the DATA directory processing block were needed. The ex-script is now functional in removing directories older than the retention period.

* Update exrtma3d_clean.ksh

Added additional conditional blocks that check for the existence of COM and DATA directories before removing individual files coinciding with the retention period.

* Update exrtma3d_clean.ksh

Added additional log file messages that highlight which cycle is being scrubbed or, if a cycle in the COM or DATA directory does not exist because it was already scrubbed or the parallel has not run long enough, a log file message is also produced.

* Update exrtma3d_clean.ksh

Re-introduced conditional block that removes empty COM directories.

---------

Co-authored-by: edward colon <edward.colon@dlogin06.dogwood.wcoss2.ncep.noaa.gov>